hugging_face/requirements.txt
peft
transformers
accelerate
tiktoken
einops
transformers_stream_generator
scipy
torchvision
pillow
tensorboard
matplotlib